% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charite_palettes.R
\docType{data}
\name{charite_palettes}
\alias{charite_palettes}
\title{Charite color palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
charite_palettes
}
\description{
A named list of color palettes conforming to the Charite visual identity.
These palettes can be used with \code{\link[=make_charite_palette]{make_charite_palette()}} and
ggplot2 scale helpers.
}
\keyword{datasets}
