% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_charite.R
\docType{data}
\name{colors_charite}
\alias{colors_charite}
\title{Color definitions in hex code for Charite - Universitaetsmedizin Berlin by Johannes Julius Mohn}
\format{
An object of class \code{list} of length 27.
}
\usage{
colors_charite
}
\value{
A hex color scheme of the official colors of Charite - Universitaetsmedizin Berlin.
}
\description{
A list of hex color codes with the primary, secondary, and accent colors for the Charite brand
as created and defined by the Charite Markenportal: https://marke.charite.de/d/Y3FxSwD6Tz3a
}
\keyword{datasets}
