% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_charite_palette.R
\name{make_charite_palette}
\alias{make_charite_palette}
\title{Interpolation or reversion of a Charite color palette}
\usage{
make_charite_palette(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in `\link{charite_palettes}}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}}
}
\value{
A function that takes an integer \code{n} and returns \code{n} interpolated colors
}
\description{
Internal function for the use of color palettes for continuous data or as many
categories as required.
}
\examples{
pal <- make_charite_palette("buch")
pal(5)
pal_rev <- make_charite_palette("mitte", reverse = TRUE)
pal_rev(10)
}
