% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_charite_palettes.R
\name{preview_charite_palettes}
\alias{preview_charite_palettes}
\title{Preview available Charite color palettes}
\usage{
preview_charite_palettes()
}
\description{
Displays all color palettes defined in \link{charite_palettes} as color swatches.
}
\examples{
preview_charite_palettes()
}
