% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_charite.R
\name{scale_color_charite}
\alias{scale_color_charite}
\title{Charite color scale for ggplot2}
\usage{
scale_color_charite(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette in \link{charite_palettes}}

\item{discrete}{Logical. Should the palette use discrete color steps?}

\item{reverse}{Logical. Should the palette be reversed?}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
A ggplot2 scale object
}
\description{
Applies a ggplot2 color scale using one of the custom color palettes from \link{charite_palettes}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, cyl, color = mpg)) +
  geom_point(size = 3) +
  scale_color_charite("berryseason", discrete = FALSE) +
  theme_charite()

ggplot(mtcars, aes(factor(cyl), color = factor(cyl))) +
  geom_point(size = 5, stat = "count") +
  scale_color_charite("secondary") +
  theme_charite()
}
