% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_charite.R
\name{scale_fill_charite}
\alias{scale_fill_charite}
\title{Charite fill scale for ggplot2}
\usage{
scale_fill_charite(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette in \link{charite_palettes}}

\item{discrete}{Logical. Should the palette use discrete color steps?}

\item{reverse}{Logical. Should the palette be reversed?}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\value{
A ggplot2 scale object
}
\description{
Applies a ggplot2 fill scale using one of the custom color palettes from \link{charite_palettes}
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_charite("goldelse") +
  theme_charite()
}
