% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sci.R
\name{theme_sci}
\alias{theme_sci}
\title{A custom ggplot2 theme for scientific publications by Johannes Julius Mohn}
\usage{
theme_sci(
  font_size = 8,
  thickness = 0.5,
  font = "sans",
  grid = FALSE,
  aspect_ratio = 1,
  tiny_margins = FALSE,
  legend_key_size = NULL,
  debug_area = FALSE
)
}
\arguments{
\item{font_size}{Base font size in points like in your favourite text editor.}

\item{thickness}{Line thickness for axes. Default ggplot2 setting is 0.5.}

\item{font}{Character string for font family. Defaults to "sans".}

\item{grid}{Logical; if TRUE, grid lines are added.}

\item{aspect_ratio}{Aspect ratio of the plot panel (y:x axis). Default is 1:1.}

\item{tiny_margins}{Logical; if TRUE, margins for all objects around the plot panel are minimized.
Font rendering may depend on your system and the output format (PDF, PNG, SVG).}

\item{legend_key_size}{Size of the legend keys. Scales with font_size by default.}

\item{debug_area}{Visualize the plot margins for plot size debugging. Defaults to FALSE.}
}
\value{
A ggplot2 theme
}
\description{
This function creates a custom ggplot2 theme with customizable options for font, line thickness, grid,
plot panel aspect ratio, and margins.
}
\details{
To use a specific font, it must be installed on your operating system.
}
\examples{
library(ggplot2)
p <- ggplot(ChickWeight, aes(x = Time, y = weight)) +
       geom_point() + theme_sci()
}
