% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virchow.R
\name{virchow}
\alias{virchow}
\title{Print ASCII art of Virchow to the console}
\usage{
virchow()
}
\value{
A fun little Easter egg printing a classic ASCII rendering of Rudolf Virchow.
}
\description{
Prints Rudolf Virchow as ASCII art.
}
