% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_plot.R
\name{example_plot}
\alias{example_plot}
\title{Generate an example plot with \code{\link[=theme_charite]{theme_charite()}}.}
\usage{
example_plot()
}
\value{
A ggplot object.
}
\description{
Creates a stylized ggplot2 plot using the built-in \link{curves} dataset to demonstrate the custom
ggplot2 Charite theme and color palettes.
}
\examples{
# generate and print the example plot
example_plot()
}
