% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_save.R
\name{nice_save}
\alias{nice_save}
\title{High-resolution (300dpi) figure export using defaults that are great
for slides and publications}
\usage{
nice_save(filename, plot = ggplot2::last_plot(), width = 3.5, ...)
}
\arguments{
\item{filename}{Name of the file to save. Should end in \code{.png}, \code{.svg}, etc.}

\item{plot}{A ggplot object. If omitted, uses the last plot.}

\item{width}{Width in inches. Default is 3.5 (ideal for slides).}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
Saves the plot and returns the filename (invisibly).
}
\description{
A wrapper around \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} that saves plots with recommended
dimensions, resolution, and format for slides and publications. By default,
saves as 3.5 inch figure, which is the standard size of one column in a
scientific publication ("full width figure"). Also consider "half width",
which measures 1.8 inch.
}
\details{
If saving to \code{.svg}, this function requires the \code{svglite} package to be installed.
You can install it with \code{install.packages("svglite")}.
}
\examples{
\dontrun{
nice_save("example.png", example_plot())
}
}
