% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_charite_palettes.R
\name{preview_charite_palettes}
\alias{preview_charite_palettes}
\title{Preview available color palettes}
\usage{
preview_charite_palettes(type = c("discrete", "continuous"))
}
\arguments{
\item{type}{Character; either "discrete" or "continuous". Defaults to "discrete".}
}
\description{
Displays all color palettes defined in \link{charite_palettes} as color swatches.
Toggle between viewing the palettes as discrete colors or continuous gradients.
}
\examples{
preview_charite_palettes()
preview_charite_palettes(type = "continuous")
}
