% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_charite.R
\name{theme_charite}
\alias{theme_charite}
\title{Custom ggplot2 theme for Charite - Universitaetsmedizin Berlin by Johannes Julius Mohn}
\usage{
theme_charite(
  base_size = 10,
  thickness = 0.5,
  grid = FALSE,
  use_calibri = FALSE
)
}
\arguments{
\item{base_size}{Base font size in points, i.e., same as in your text editor.}

\item{thickness}{Line thickness for axes. Default ggplot2 setting is 0.5.}

\item{grid}{Logical; if TRUE, grid lines are added.}

\item{use_calibri}{Use Calibri font? Defaults to FALSE.}
}
\value{
A ggplot2 theme by Johannes Julius Mohn for Charite - Universitaetsmedizin Berlin.
}
\description{
This function creates a custom ggplot2 theme with customizable options for font, line thickness, and grid.
You can optionally choose to plot text in Calibri font if available on your OS.
Calibri is the standard font on Windows and recommended by Charite. On macOS and Linux you may need to
manually install the font first.
}
