% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virchow.R
\name{virchow}
\alias{virchow}
\title{ASCII art of Rudolf Virchow}
\usage{
virchow()
}
\value{
A fun little Easter egg printing an ASCII rendering of Rudolf Virchow.
}
\description{
Prints Rudolf Virchow as ASCII art to the R console.
}
